Move filter dropdown back to game dashboard
